+++ test/plugins/elixir/doctest-io-fixtures/in_string.ex
@@ -1,0 +1,4 @@
+defmodule InString do
+  @msg "For now, this string is OK"
+  def msg, do: @msg
+end