+++ test/plugins/elixir/doctest-io-fixtures/untagged.ex
@@ -1,0 +1,4 @@
+defmodule Untagged do
+  # For now, hardcoded timeout
+  @timeout 5000
+end